requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
autoflake==2.3.3
autopep8==2.3.2
bcrypt==5.0.0
black==26.3.0
blinker==1.9.0
build==1.4.0
cachetools==7.0.4
certifi==2026.2.25
cfgv==3.5.0
charset-normalizer==3.4.5
chromadb==1.5.4
click==8.3.1
colorama==0.4.6
distlib==0.4.0
distro==1.9.0
durationpy==0.10
et_xmlfile==2.0.0
filelock==3.25.0
flake8==7.1.1
flatbuffers==25.12.19
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.73.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.6.0
identify==2.6.17
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
isort==8.0.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==2.17.0
networkx==3.6.1
nltk==3.9.3
nodeenv==1.10.0
numpy==2.4.3
oauthlib==3.3.1
onnxruntime==1.24.3
openai==2.26.0
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
pre_commit==4.5.1
protobuf==6.33.5
pyarrow==23.0.1
pybase64==1.4.3
pycodestyle==2.12.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.19.2
PyMuPDF==1.27.1
PyPika==0.51.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
pytest==9.0.2
python-dateutil==2.9.0.post0
python-discovery==1.1.3
python-docx==1.2.0
python-dotenv==1.2.2
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
virtualenv==21.2.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
zipp==3.23.0
